Redesigned the IO layer to only utilize a single thread/selector.  Unfortunately, having multiple selectors (although a good load sharing idea in theory) has the serious potential to lead to deadlock and race conditions when the multiple selector threads start accessing shared SocketChannels and Keys.  This could easily be fixed by using synchronized(){} statements, but then the whole point of multi-selectors is lost as the selectors are now waiting on resources anyways....
ConnectionManager is a singleton patterned class.  It houses a map of SocketChannels to Connection objects.  It also houses the selector and all IO routines.  It handles new connections and disconnections.  It implements Runnable and has a self contained thread.  Thread controls are delegated out to the ConnectionManager object and publicly accessible. The ConnectionManager.run() contains a loop controllable by .start() and .stop().  These loop controls do not erase object parameters and thus .start() may be called even after a .stop() has been called.

Further refinement of this class is warranted to allow for better Thread management and to add the ability to prevent any more incoming connections at the Socket level.