Added the ground work for the Job handling system.
JobManager is a singleton patterned class.  It contains the queue of jobs to be executed by n number of JobWorker objects.

JobWorker objects will implement Runnable and will be controlled by JobManager.

AbstractJob class provides the base mandatory implementation of a Job.  ALL jobs are to extend AbstractJob.  (Consider making this fully virtual)

CheckConnBufferForValidCmd_Job class:  A job created by the ConnectionManager to test to see if the contained Connection object's readBuffer contains a complete jMUD command yet.  If it does, then it is parsed out of the ByteBuffer and a subsequent BuildCmdFromString_Job is generated.  If not, then the Connection object's readBuffer is left untouched.

BuildCmdFromString_Job class:  A Job created to see if a jMUD command can be built into a command object.  This is accomplished by checking the String to jMudCommand map in the CommandManager.