I forgot to mention it (was very tired) but the commit that is r58 was the start of the segregation between jmud framework and a mud implementation.
jmud.engine.* is where all framework components will reside.  Only things that are Abstract and/or essential to the operation of the jmud server are to be stored here. jmud.engine.* should be compilable AND runnable as a "barebones" server.  

jmud.testserver.* is where any MUD specific code will go, aka, all the extentions from jmud.engine.commands.AbstractCommandDef.