Laid down framework for configuring the jmud: AbstractConfig and ConfigManager
In short, since nearly all the *Managers in jmud rely on single instances of Definition classes to be registered with a Manager/Registrar (StatDef, SlotDef, etc) an easy way is needed to instantiate these instances and load them into the manager.  The config system will do that by providing a Base AbstractConfig that the mud Dev extends from and populates the .doConfig() method with the code that instantiates and registers their Def classes.  The ConfigManager.executeConfigs() method is called at the very startup of the mud.